requirements.txt
Django>=2.2.0,<2.3.0
djangorestframework>=3.9.4,<4.0.0
Pillow>=6.0.0,<7.0.0
django-import-export>=1.2.0,<1.3.0
xhtml2pdf>=0.2.3,<0.3.0
django-easy-audit>=1.1.0,<1.5.0
django-debug-toolbar==2.0
flake8>=3.7.7,<3.9.0
django-background-tasks>=1.2.0,<1.5.0
coveralls>=1.8.1,<2.0.0
requests>=2.2.0,<3.0.0
psycopg2-binary>=2.8.3,<3.0.0
django-allauth>=0.39.1,<0.50.0
django-rest-auth>=0.9.5,<1.0.0
python-keycloak==0.12.0
python-jose==3.0.0

